Split the media/ GET operation in a read-only controller
